Struct massa_hash::hash::Hash

pub struct Hash(Hash);

Hash wrapper, the underlying hash type is Blake3

The motivations for selecting Blake3 were- Speed: Blake3 is significantly faster than other popular hashing algorithms, such as SHA-256 and SHA-3. This is largely due to its ability to leverage modern CPU architectures and instruction sets, as well as its optimized implementation.

Security: Blake3 is designed to be highly secure and resistant to a wide range of attacks, including collision attacks, length-extension attacks, and timing attacks. It also offers better resistance to side-channel attacks than many other hashing algorithms.

Flexibility: Blake3 is highly flexible and can be used in a variety of applications, including as a general-purpose hash function, as a key derivation function, and as a message authentication code. It also supports a wide range of input sizes and can produce output of any desired length. Scalability: Blake3 can efficiently take advantage of multiple cores and SIMD (single instruction, multiple data) instructions, allowing it to scale well on modern CPUs.

Improved Compression Function: The compression function used in Blake3 is an improved version of the one used in its predecessor, Blake2. This improved compression function offers better diffusion and mixing properties, which contributes to its increased security.

Keyed Hashing: Blake3 supports keyed hashing, which allows users to use a secret key to generate a unique hash value. This feature can be useful in applications that require message authentication or integrity verification.

Tree Hashing: Blake3 supports tree hashing, which allows users to hash large files or data structures in a parallel and efficient manner. This feature can be useful in applications that involve large-scale data processing, such as cloud storage or distributed file systems.

Open Source: Blake3 is an open-source algorithm, which means that its code is publicly available for review and auditing by anyone. This can help improve its security and reliability, as well as increase transparency and trust among users.

Tuple Fields
0: Hash
Implementations

impl Hash

pub fn zero() -> Self

Creates a hash full of zeros bytes.

pub fn compute_from(data: &[u8]) -> Self

Compute a hash from data.

Example

let hash = Hash::compute_from(&"hello world".as_bytes());

pub fn compute_from_tuple(data: &[&[u8]]) -> Self

Compute a hash from tuple of byte arrays.

Example

let hash = Hash::compute_from_tuple(&[&"hello".as_bytes(), &"world".as_bytes()]);

pub fn to_bs58_check(&self) -> String

Serialize a Hash using bs58 encoding with checksum.

Example

let hash = Hash::compute_from(&"hello world".as_bytes());
let serialized: String = hash.to_bs58_check();

Motivations for using base58 encoding:

base58_check is like base64 but-

fully standardized (no = vs /)

no weird characters (eg. +) only alphanumeric

ambiguous letters combined (eg. O vs 0, or l vs 1)

contains a checksum at the end to detect typing errors

pub fn to_bytes(&self) -> &[u8; 32]

Serialize a Hash as bytes.

Example

let hash = Hash::compute_from(&"hello world".as_bytes());
let serialized = hash.to_bytes();

pub fn into_bytes(self) -> [u8; 32]

Convert into bytes.

Example

let hash = Hash::compute_from(&"hello world".as_bytes());
let serialized = hash.into_bytes();

pub fn from_bs58_check(data: &str) -> Result<Hash, MassaHashError>

Deserialize using bs58 encoding with checksum.

Example

let hash = Hash::compute_from(&"hello world".as_bytes());
let serialized: String = hash.to_bs58_check();
let deserialized: Hash = Hash::from_bs58_check(&serialized).unwrap();

pub fn from_bytes(data: &[u8; 32]) -> Hash

Deserialize a Hash as bytes.

Example

let hash = Hash::compute_from(&"hello world".as_bytes());
let serialized = hash.into_bytes();
let deserialized: Hash = Hash::from_bytes(&serialized);